guix: giqt declaration now uses cl-gobject-introspection input
This ensure guix giqt environment is repeatable.

Note: this leaves some dead code were quicklisp was previously used. For
systems like debian/ubuntu or other distro's where cl-gobject-instrospection
might not be installed. This dependency is the left to the user to install.

The dead quicklisp dead code is left as a convenience for me until I can
switch to asdf.