guix: g-golf package now substitutes absolute path to libg-golf
Because lib-golf is needed during building. It can not be substituted before
compilation. This patch substitutes it after and then recompile the (g-golf
init) modules.

This also cleans up the re basing of the modules. Instead of copying the
modules we now just re base by moving them. This way we do not leave orphaned
files behind.