guix: g-golf package now substitutes libg-gofl before build phase
This adds the environment variable GUILE_GGOLF_UNINSTALLED when set to use the
default dynamic lib locations. If not it uses the absolute path in the guix
store.

The g-golf package now substitutes configure to install the guix guile modules
into a re based guile site path. Which avoids having to move the modules after
installing.

g-golf is now at revision 574 and it which now supports GTK signals.